% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog.R
\name{flog}
\alias{flog}
\title{Convenient wrapper around logging functionality from the futile.logger package}
\usage{
flog(..., level = c("info", "debug", "warn"))
}
\arguments{
\item{...}{Concatenated to form a message}

\item{level}{Logging level to use}
}
\description{
Convenient wrapper around logging functionality from the futile.logger package
}
